requirements.txt
tensorflow==2.20.0
flask==3.1.2
pillow==11.3.0
numpy==2.0.2
flask-login==0.6.3
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.1
